requirements.txt
# Core Libraries
chromadb==0.4.14                 # Vector database for embeddings
sentence-transformers==2.2.2     # Generating text embeddings
ollama==0.1.6                    # Local LLM integration

# PDF and Document Parsing
PyMuPDF==1.23.3                  # PDF parsing library
python-docx==1.1.0               # Parsing .docx files (optional)

# Image Processing and OCR
pytesseract==0.3.10              # OCR for extracting text from images
Pillow==10.0.0                   # Image handling library

# Natural Language Processing
nltk==3.8.1                      # NLP tools for tokenization, etc.

# Data Manipulation & Utilities
numpy==1.25.2                    # Numerical operations
scipy==1.11.2                    # Scientific computations

# Logging and Debugging
loguru==0.7.0                    # Advanced logging library

# Environment Variable Management
python-dotenv==1.0.0             # Load environment variables from .env

# Progress Bars and CLI Tools
tqdm==4.66.1                     # For showing progress bars during processing
click==8.1.7                     # Building interactive command-line interfaces

# JSON and YAML Support (Optional)
PyYAML==6.0.1                    # Parsing YAML files
